refactor: extract PowerShellFunctions from NABFunctions
